Update docstring of gbasis/contractions to maintain consistent notation
Notation changes:
K is used solely as the number of contractions (N_contractions) in each CGTO.
Modified instances where L was used as the number of contracted Cartesian Gaussian
primitives, to avoid confusing the number of primitives with the angular momentum.
Clarified that num_contr reflects the number of Cartesian primitives of a given
angular momentum.